requirements.txt
faker==4.1.0
ruamel-yaml==0.16.10
requests==2.23.0
pymysql==0.9.3
paramiko==2.7.1
configparser==5.0.0
locust==1.0.1
numpy==1.18.4
pytest==5.4.3
pytest-ordering==0.6
pytest-html==2.1.1
allure-pytest==2.8.18
pytest-rerunfailures==9.1.1
pytest-ordering==0.6
